Modify reduce function to reuse _.each function
One of the tests in _.contains function ('should return true given a object and a value from that object') failed. Since the _.contains function used the _.reduce function, problem was in the code for _.reduce, specifically involving iteration over objects. By modifying the _.reduce function to use _.each, iteration over objects was accounted for and thus made all the tests pass for the _.contains function.